Major update Data Mitra, Data Pengunjung, Surat masuk and keluar. Add Data Pengurus

+++ admin/proses/proses_editsuratkeluar.php
@@ -2,14 +2,15 @@
 
 include '../../koneksi/koneksi.php';
 
-$id = $_POST['id'];
+$No = $_POST['No'];
 $tgl_keluar = $_POST['tanggal_keluar'];
 $kode = $_POST['kode'];
 $nomor_surat = $_POST['nomor_surat'];
 $penerima = $_POST['penerima'];
 $perihal = $_POST['perihal'];
+$keterangan = $_POST['keterangan'];
 
-$query_get_file = "SELECT file_surat FROM tb_arsip_surat_keluar WHERE id = $id";
+$query_get_file = "SELECT file_surat FROM tb_arsip_surat_keluar WHERE No = $No";
 $result = mysqli_query($db, $query_get_file);
 $row = mysqli_fetch_assoc($result);
 $file_lama = $row['file_surat'];
@@ -22,19 +23,29 @@ if (isset($_FILES['file_surat'])&&$_FILES["file_surat"]["error"] === UPLOAD_ERR_
 
     // Path untuk menyimpan file
     $target_dir = "../uploads/";
-    $destination = $target_dir . $nama_file; if (move_uploaded_file($_FILES['file_surat']['tmp_name'], $destination)) {
+    $destination = $target_dir . $nama_file;
+    if (move_uploaded_file($_FILES['file_surat']['tmp_name'], $destination)) {
         if (file_exists($file_lama)) {
             unlink($file_lama);
         }
         // Jika file baru berhasil dipindahkan, update nama file di database
-        $query = "UPDATE tb_arsip_surat_keluar SET tanggal_keluar='$tgl_keluar', nomor_surat='$nomor_surat', penerima='$penerima', perihal='$perihal', kode='$kode' , file_surat='$destination' WHERE id='$id'";
+        $query = "UPDATE tb_arsip_surat_keluar SET 
+                  No = '$No',
+                  tanggal_keluar = '$tgl_keluar',
+                  nomor_surat = '$nomor_surat',
+                  penerima = '$penerima',
+                  perihal = '$perihal',
+                  kode = '$kode',
+                  keterangan = '$keterangan',
+                  file_surat = '$destination'
+                  WHERE id = '$id'";
     } else {
         echo "Gagal memindahkan file.";
         exit;
     }
 } else {
     // Jika tidak ada file baru, tetap gunakan file lama
-    $query = "UPDATE tb_arsip_surat_keluar SET tanggal_keluar='$tgl_keluar', nomor_surat='$nomor_surat', penerima='$penerima', perihal='$perihal', kode='$kode' WHERE id='$id'";
+    $query = "UPDATE tb_arsip_surat_keluar SET tanggal_keluar='$tgl_keluar', nomor_surat='$nomor_surat', penerima='$penerima', perihal='$perihal', kode='$kode' WHERE No='$No'";
 }
 
 if (mysqli_query($db, $query)) {
